Rewrite all tests to work with new system. Remove redundant variables.
